renamed class _Convert.cs to MyConvert.cs and used LINQ functionality to minimize code
