--- .idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]1/Sorter.csproj.CoreCompileInputs.cache
@@ -1,1 +1,0 @@
-b09f7fc0f69664b58bf784ead925bd37561f5e78
