--- .idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47__Default_Changelist_.xml
@@ -1,79 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]" date="1618483647670" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 15.04.2021 13:47 [Default Changelist]" />
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.csprojAssemblyReference.cache" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.csprojAssemblyReference.cache" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter.csprojAssemblyReference.cache" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/LanguageWindow.baml" />
-    <option name="AFTER_PATH" />
-    <option name="SHELVED_PATH" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/MenuWindow.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/MenuWindow.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/MenuWindow.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.csproj.CoreCompileInputs.cache" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.csproj.CoreCompileInputs.cache" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter.csproj.CoreCompileInputs.cache" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/ref/Sorter.dll" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/ref/Sorter.dll" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter.dll" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/res/dimens.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/res/dimens.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/dimens.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.dll" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.dll" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter1.dll" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/App.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/App.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/App.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/res/colors.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/res/colors.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/colors.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/bin/Debug/net5.0-windows/ref/Sorter.dll" />
-    <option name="AFTER_PATH" value="Sorter/bin/Debug/net5.0-windows/ref/Sorter.dll" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter2.dll" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.g.resources" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.g.resources" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter.g.resources" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.pdb" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.pdb" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter.pdb" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/res/style.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/res/style.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/style.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/bin/Debug/net5.0-windows/Sorter.pdb" />
-    <option name="AFTER_PATH" value="Sorter/bin/Debug/net5.0-windows/Sorter.pdb" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter1.pdb" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/bin/Debug/net5.0-windows/Sorter.dll" />
-    <option name="AFTER_PATH" value="Sorter/bin/Debug/net5.0-windows/Sorter.dll" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Checkout_at_15_04_2021_13_47_[Default_Changelist]/Sorter3.dll" />
-  </binary>
-</changelist>
