--- .idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/Sorter_MarkupCompile.cache
@@ -1,20 +1,0 @@
-﻿Sorter
-
-
-winexe
-C#
-.cs
-C:\Users\maxhe\RiderProjects\Sorter\Sorter\obj\Debug\net5.0-windows\
-Sorter
-none
-false
-TRACE;DEBUG;NET;NET5_0;NETCOREAPP
-C:\Users\maxhe\RiderProjects\Sorter\Sorter\App.xaml
-5-187946475
-
-7-1317664914
-193-1891726104
-LanguageWindow.xaml;MenuWindow.xaml;res\colors.xaml;res\dimens.xaml;res\style.xaml;
-
-True
-
