--- .idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48__Default_Changelist_.xml
@@ -1,44 +1,0 @@
-<changelist name="Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]" date="1618483746593" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Merge at 15.04.2021 13:48 [Default Changelist]" />
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/MenuWindow.baml" />
-    <option name="AFTER_PATH" />
-    <option name="SHELVED_PATH" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/App.baml" />
-    <option name="AFTER_PATH" />
-    <option name="SHELVED_PATH" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter_MarkupCompile.cache" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter_MarkupCompile.cache" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/Sorter_MarkupCompile.cache" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/res/dimens.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/res/dimens.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/dimens.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/res/style.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/res/style.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/style.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/LanguageWindow.baml" />
-    <option name="AFTER_PATH" />
-    <option name="SHELVED_PATH" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/res/colors.baml" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/res/colors.baml" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/colors.baml" />
-  </binary>
-  <binary>
-    <option name="BEFORE_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.csprojAssemblyReference.cache" />
-    <option name="AFTER_PATH" value="Sorter/obj/Debug/net5.0-windows/Sorter.csprojAssemblyReference.cache" />
-    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/.idea.Sorter/.idea/shelf/Uncommitted_changes_before_Merge_at_15_04_2021_13_48_[Default_Changelist]/Sorter.csprojAssemblyReference.cache" />
-  </binary>
-</changelist>
